+++ cmd/get/launch_plan.go
@@ -2,38 +2,12 @@ package get
 
 import (
 	"context"
-	"encoding/json"
 	"github.com/lyft/flytectl/cmd/config"
 	cmdCore "github.com/lyft/flytectl/cmd/core"
 	"github.com/lyft/flytectl/pkg/printer"
 	"github.com/lyft/flyteidl/gen/pb-go/flyteidl/admin"
 	"github.com/lyft/flytestdlib/logger"
 )
-
-// PrintableLaunchPlan is the structure for printing Launch Plan
-type PrintableLaunchPlan struct {
-	Version          string `header:"Version"`
-	Name             string `header:"Name"`
-	Type             string `header:"Type"`
-	Discoverable     bool   `header:"Discoverable"`
-	DiscoveryVersion string `header:"DiscoveryVersion"`
-}
-
-var launchPlanStructure = map[string]string{
-	"Version":          "$.id.version",
-	"Name":             "$.id.name",
-	"Type":             "$.closure.compiledTask.template.type",
-	"Discoverable":     "$.closure.compiledTask.template.metadata.discoverable",
-	"DiscoveryVersion": "$.closure.compiledTask.template.metadata.discovery_version",
-}
-
-func transformLaunchPlan(jsonbody []byte) (interface{}, error) {
-	results := PrintableLaunchPlan{}
-	if err := json.Unmarshal(jsonbody, &results); err != nil {
-		return results, err
-	}
-	return results, nil
-}
 
 func getLaunchPlanFunc(ctx context.Context, args []string, cmdCtx cmdCore.CommandContext) error {
 	launchPlanPrinter := printer.Printer{}
@@ -48,13 +22,12 @@ func getLaunchPlanFunc(ctx context.Context, args []string, cmdCtx cmdCore.Comman
 	if err != nil {
 		return err
 	}
-
 	if len(args) == 1 {
 		name := args[0]
 		logger.Debugf(ctx, "Retrieved %v excutions", len(launchPlans.LaunchPlans))
 		for _, v := range launchPlans.LaunchPlans {
 			if v.Id.Name == name {
-				err := launchPlanPrinter.Print(config.GetConfig().MustOutputFormat(), launchPlans, launchPlanStructure, transformLaunchPlan)
+				err := launchPlanPrinter.Print(config.GetConfig().MustOutputFormat(), v, launchPlanStructure, transformLaunchPlan)
 				if err != nil {
 					return err
 				}
@@ -63,8 +36,7 @@ func getLaunchPlanFunc(ctx context.Context, args []string, cmdCtx cmdCore.Comman
 		}
 		return nil
 	}
-
 	logger.Debugf(ctx, "Retrieved %v launch plan", len(launchPlans.LaunchPlans))
-	return launchPlanPrinter.Print(config.GetConfig().MustOutputFormat(), launchPlans, launchPlanStructure, transformLaunchPlan)
+	return launchPlanPrinter.Print(config.GetConfig().MustOutputFormat(), launchPlans.LaunchPlans, launchPlanStructure, transformLaunchPlan)
 	return nil
 }
