+++ cmd/get/project.go
@@ -2,7 +2,6 @@ package get
 
 import (
 	"context"
-	"encoding/json"
 
 	"github.com/lyft/flyteidl/gen/pb-go/flyteidl/admin"
 	"github.com/lyft/flytestdlib/logger"
@@ -11,27 +10,6 @@ import (
 	cmdCore "github.com/lyft/flytectl/cmd/core"
 	"github.com/lyft/flytectl/pkg/printer"
 )
-
-// PrintableProject is the structure for printing Project
-type PrintableProject struct {
-	ID          string `header:"Id"`
-	Name        string `header:"Name"`
-	Description string `header:"Description"`
-}
-
-var tableStructure = map[string]string{
-	"ID":          "$.id",
-	"Name":        "$.name",
-	"Description": "$.description",
-}
-
-func transformProject(jsonbody []byte) (interface{}, error) {
-	results := PrintableProject{}
-	if err := json.Unmarshal(jsonbody, &results); err != nil {
-		return results, err
-	}
-	return results, nil
-}
 
 func getProjectsFunc(ctx context.Context, args []string, cmdCtx cmdCore.CommandContext) error {
 	adminPrinter := printer.Printer{}
