+++ cmd/get/workflow.go
@@ -2,7 +2,6 @@ package get
 
 import (
 	"context"
-	"encoding/json"
 
 	"github.com/lyft/flytestdlib/logger"
 
@@ -13,25 +12,6 @@ import (
 
 	"github.com/lyft/flyteidl/gen/pb-go/flyteidl/admin"
 )
-
-var workflowStructure = map[string]string{
-	"Version": "$.id.version",
-	"Name":    "$.id.name",
-}
-
-// PrintableWorkflow is the structure for printing workflow
-type PrintableWorkflow struct {
-	Name    string `header:"Name"`
-	Version string `header:"Version"`
-}
-
-var transformWorkflow = func(jsonbody []byte) (interface{}, error) {
-	results := PrintableWorkflow{}
-	if err := json.Unmarshal(jsonbody, &results); err != nil {
-		return results, err
-	}
-	return results, nil
-}
 
 func getWorkflowFunc(ctx context.Context, args []string, cmdCtx cmdCore.CommandContext) error {
 	adminPrinter := printer.Printer{}
